third_party/blink/web_tests/fast/css/clear-activechain-list-shadow-dom-expected.txt

Click Me

Chain of active elements should be cleared including the Shadow DOM elements

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS background is "rgb(0, 255, 0)"
PASS background is "rgb(255, 0, 0)"
PASS background is "rgb(0, 255, 0)"
PASS successfullyParsed is true

TEST COMPLETE
